[ :recycle: ] | refact: rename (module::application)
