+++ tutorialls/src/application/users/users.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from './users.controller';
-import { AuthService } from './users.service';
+import { AuthService } from './auth.service';
 
 describe('UsersController', () => {
   let controller: UsersController;
