+++ tutorialls/src/domain/repository/user/user.repository.ts
@@ -4,7 +4,7 @@ import { IFindUserByIdDTO } from 'src/domain/DTO/user/find/by/id.dto';
 import { User } from 'src/domain/entity/user.entity';
 
 export interface IUserRepository {
-  register(user: ISignupUserDTO): Promise<User>;
+  signup(user: ISignupUserDTO): Promise<User>;
   findByEmail(DTO: IFindUserByEmailDTO): Promise<User>;
   findById(DTO: IFindUserByIdDTO): Promise<User>;
 }
